backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
kagglehub==0.2.2
pandas==2.2.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.37.2
torch
python-dotenv==1.0.1
requests
scikit-learn
